Refactorized

+++ database/dbmodel.py
@@ -21,7 +21,7 @@ class Pool(db.Model):
     ID = db.Column(db.String(80), primary_key=True)
     Name = db.Column(db.String(80), nullable=False)
     MaximumCount = db.Column(db.Integer)
-    Description = db.Column(db.String(80))
+    Description = db.Column(db.String(200))
     Enabled = db.Column(db.Boolean)
     OSID = db.Column(db.Integer, db.ForeignKey("OperatingSystem.ID"))
     Software = db.relationship("SoftwareList")
@@ -50,21 +50,47 @@ class Pool(db.Model):
     @staticmethod
     def rm_pool(pool_id):
         pool = Pool.query.filter(Pool.ID == pool_id).first()
+        if not pool:
+            return False
         software_list = SoftwareList.query.filter(SoftwareList.PoolID == pool_id).all()
         for sw in software_list:
             db.session.delete(sw)
         db.session.delete(pool)
         db.session.commit()
+        return True
 
     @staticmethod
-    def edit_pool(pool_id, id, name, max_count, description, enabled):
+    def edit_pool(pool_id, new_id, name, max_count, description, enabled):
+        if pool_id != new_id and (new_id,) in Pool.query.filter().with_entities(Pool.ID).all():
+            return False
         pool = Pool.query.filter(Pool.ID == pool_id).first()
-        pool.ID = id
+        pool.ID = new_id
         pool.Name = name
         pool.MaximumCount = max_count
         pool.Description = description
         pool.Enabled = enabled
         db.session.commit()
+
+        for sw in SoftwareList.query.filter(SoftwareList.PoolID == pool_id).all():
+            sw.PoolID = new_id
+            db.session.commit()
+
+        return True
+
+    @staticmethod
+    def edit_software(pool_id, new_software_list):
+        pool = Pool.query.filter(Pool.ID == pool_id).first()
+        old_software_list = {tuple(x[1:]) for x in pool.get_software_list()}
+        new_software_list = {tuple(x) for x in new_software_list}
+
+        software_to_remove = old_software_list.difference(new_software_list)
+        for name, version in software_to_remove:
+            pool.remove_software(Software.get_software_by_name(name), version)
+
+        software_to_add = new_software_list.difference(old_software_list)
+        for name, version in software_to_add:
+            software = Software.add_software(name)
+            pool.add_software(software, version)
 
     def get_software_list(self, software=None):
         if software is None:
@@ -249,6 +275,10 @@ class Software(db.Model):
     @staticmethod
     def get_software(software_id):
         return Software.query.filter(Software.ID == software_id).first()
+
+    @staticmethod
+    def get_software_by_name(name):
+        return Software.query.filter(Software.Name == name).first()
 
     @staticmethod
     def add_software(software_name):
